requirements.txt
apted>=1.0.0
lxml>=4.9.0
python-Levenshtein>=0.21.0

